move queries
