+++ packages/core/package.json
@@ -14,6 +14,7 @@
     "./member-profile.server": "./src/member-profile.server.ts",
     "./member-profile.ui": "./src/member-profile.ui.ts",
     "./admin.server": "./src/modules/admin/index.server.ts",
+    "./applications": "./src/modules/application/application.core.ts",
     "./education": "./src/modules/education/education.core.ts",
     "./education.types": "./src/modules/education/education.types.ts",
     "./education.ui": "./src/modules/education/education.ui.tsx",
