+++ packages/core/src/admin-dashboard.server.ts
@@ -2,9 +2,6 @@ export { QueueFromName } from './infrastructure/bull/bull';
 export { job } from './infrastructure/bull/use-cases/job';
 export { listAdmins } from './modules/admin/queries/list-admins';
 export { addAdmin } from './modules/admin/use-cases/add-admin';
-export { countPendingApplications } from './modules/application/queries/count-pending-applications';
-export { getApplication } from './modules/application/queries/get-application';
-export { listApplications } from './modules/application/queries/list-applications';
 export { acceptApplication } from './modules/application/use-cases/accept-application';
 export { rejectApplication } from './modules/application/use-cases/reject-application';
 export { updateEmailApplication } from './modules/application/use-cases/update-application-email';
