+++ packages/core/src/modules/application/application.core.ts
@@ -1,7 +1,9 @@
 import dayjs from 'dayjs';
+import { type SelectExpression, sql } from 'kysely';
 import { match } from 'ts-pattern';
 
-import { db } from '@oyster/db';
+import { db, type DB } from '@oyster/db';
+import { ApplicationStatus } from '@oyster/types';
 import { iife } from '@oyster/utils';
 
 import {
@@ -11,6 +13,126 @@ import {
 import { job } from '@/infrastructure/bull/use-cases/job';
 import { registerWorker } from '@/infrastructure/bull/use-cases/register-worker';
 import { reviewApplication } from './use-cases/review-application';
+
+// Queries
+
+// "Count Pending Applications"
+
+export async function countPendingApplications() {
+  const result = await db
+    .selectFrom('applications')
+    .select((eb) => eb.fn.countAll<string>().as('count'))
+    .where('status', '=', ApplicationStatus.PENDING)
+    .executeTakeFirstOrThrow();
+
+  const count = Number(result.count);
+
+  return count;
+}
+
+// "Get Application"
+
+type GetApplicationOptions = {
+  withSchool?: boolean;
+};
+
+export async function getApplication<
+  Selection extends SelectExpression<DB, 'applications'>,
+>(id: string, selections: Selection[], options: GetApplicationOptions = {}) {
+  const result = await db
+    .selectFrom('applications')
+    .select(selections)
+    .$if(!!options.withSchool, (qb) => {
+      return qb
+        .leftJoin('schools', 'schools.id', 'applications.schoolId')
+        .select(['schools.name as school']);
+    })
+    .where('applications.id', '=', id)
+    .executeTakeFirst();
+
+  return result;
+}
+
+// "List Applications"
+
+type ApplicationsSearchParams = {
+  limit: number;
+  page: number;
+  search: string;
+  status: ApplicationStatus | 'all';
+  timezone: string;
+};
+
+export async function listApplications({
+  limit,
+  page,
+  search,
+  status,
+  timezone,
+}: ApplicationsSearchParams) {
+  const query = db
+    .selectFrom('applications')
+    .$if(!!search, (qb) => {
+      return qb.where((eb) =>
+        eb.or([
+          eb('applications.email', 'ilike', `%${search}%`),
+          eb('applications.firstName', 'ilike', `%${search}%`),
+          eb('applications.lastName', 'ilike', `%${search}%`),
+          eb(
+            sql`applications.first_name || ' ' || applications.last_name`,
+            'ilike',
+            `%${search}%`
+          ),
+        ])
+      );
+    })
+    .$if(status !== 'all', (qb) => {
+      return qb.where('applications.status', '=', status);
+    });
+
+  const orderDirection = status === 'pending' ? 'asc' : 'desc';
+
+  const [rows, { count }] = await Promise.all([
+    query
+      .leftJoin('schools', 'schools.id', 'applications.schoolId')
+      .leftJoin('admins', 'admins.id', 'applications.reviewedById')
+      .select([
+        'applications.createdAt',
+        'applications.email',
+        'applications.firstName',
+        'applications.id',
+        'applications.lastName',
+        'applications.status',
+        'admins.firstName as reviewedByFirstName',
+        'admins.lastName as reviewedByLastName',
+        (eb) => {
+          return eb.fn
+            .coalesce('schools.name', 'applications.otherSchool')
+            .as('school');
+        },
+      ])
+      .orderBy('applications.createdAt', orderDirection)
+      .limit(limit)
+      .offset((page - 1) * limit)
+      .execute(),
+
+    query
+      .select((eb) => eb.fn.countAll<string>().as('count'))
+      .executeTakeFirstOrThrow(),
+  ]);
+
+  const applications = rows.map((row) => {
+    return {
+      ...row,
+      createdAt: dayjs(row.createdAt).tz(timezone).format('MM/DD/YY @ h:mm A'),
+    };
+  });
+
+  return {
+    applications,
+    totalCount: Number(count),
+  };
+}
 
 // Worker
